annotation: using dynamic flow allows skipping steps
The AnnotBatch job is built using a flow for each step.
Each flow uses a decider to skip or include the step.
The decider has a pipelineOptions to retrieve whether it must
skip the step or not.

The VariantJob still uses the old annotation steps. There's a dedicated
"variantAnnotBatchJob", and the usual skip* work with the new job as well:

skipAnnotGenerateInput=false
skipAnnotCreate=true
skipAnnotLoad=true
will only call the generateInputBatch step